# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/mobile/devices:
    get:
      operationId: listMobileDevices
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileDevicesResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Service Unavailable
      summary: List paired mobile devices with their live/muted status
      tags:
      - mobile
  /api/v1/mobile/devices/{installId}:
    delete:
      operationId: removeMobileDevice
      parameters:
      - description: The device's stable install id.
        in: path
        name: installId
        required: true
        schema:
          description: The device's stable install id.
          type: string
      responses:
        "204":
          description: No Content
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Service Unavailable
      summary: Remove a paired device from the roster
      tags:
      - mobile
    patch:
      operationId: muteMobileDevice
      parameters:
      - description: The device's stable install id.
        in: path
        name: installId
        required: true
        schema:
          description: The device's stable install id.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MuteDeviceRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: boolean
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Service Unavailable
      summary: Mute or unmute push notifications for a paired device
      tags:
      - mobile
  /api/v1/mobile/disable:
    post:
      operationId: disableMobile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Disable the Connect Mobile LAN bridge
      tags:
      - mobile
  /api/v1/mobile/enable:
    post:
      operationId: enableMobile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Enable the Connect Mobile LAN bridge and issue a fresh password
      tags:
      - mobile
  /api/v1/mobile/regenerate:
    post:
      operationId: regenerateMobile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Rotate the Connect Mobile password, dropping any connected phone
      tags:
      - mobile
  /api/v1/mobile/secure-pairing:
    post:
      operationId: setMobileSecurePairing
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControllersSetSecurePairingRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Turn TLS-over-Tailscale secure pairing on or off
      tags:
      - mobile
  /api/v1/mobile/status:
    get:
      operationId: getMobileStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
      summary: Check whether Connect Mobile's LAN bridge is enabled
      tags:
      - mobile
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    ControllersSecurePairingStatus:
      properties:
        active:
          type: boolean
        available:
          type: boolean
        enabled:
          type: boolean
        host:
          type: string
        port:
          type: integer
        reason:
          type: string
      required:
      - enabled
      - available
      - active
      - host
      - port
      - reason
      type: object
...
    ControllersSetSecurePairingRequest:
      properties:
        enabled:
          type: boolean
      required:
      - enabled
      type: object
...
    MobileDeviceResponse:
      properties:
        createdAt:
          format: date-time
          type: string
        deviceName:
          type: string
        installId:
          type: string
        lastSeenAt:
          format: date-time
          type: string
        live:
          description: True when the phone's app is open and polling.
          type: boolean
        muted:
          type: boolean
        notificationsEnabled:
          description: True when this device has a push token registered.
          type: boolean
        platform:
          enum:
          - ios
          - android
          type: string
        token:
          type: string
      required:
      - installId
      - muted
      - live
      - notificationsEnabled
      - createdAt
      - lastSeenAt
      type: object
    MobileDevicesResponse:
      properties:
        devices:
          items:
            $ref: '#/components/schemas/MobileDeviceResponse'
          type: array
      required:
      - devices
      type: object
    MobileStatusResponse:
      properties:
        enabled:
          type: boolean
        host:
          type: string
        password:
          type: string
        port:
          type: integer
        securePairing:
          $ref: '#/components/schemas/ControllersSecurePairingStatus'
        tailscaleHost:
          type: string
        warning:
          type: string
      required:
      - enabled
      - host
      - tailscaleHost
      - port
      - password
      - warning
      - securePairing
      type: object
    MuteDeviceRequest:
      properties:
        muted:
          description: True to stop sending push notifications to this device.
          type: boolean
      required:
      - muted
      type: object
...